requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
google-play-scraper
